+++ scripts/babel-relay-plugin/lib/RelayQLAST.js
@@ -16,7 +16,7 @@
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; desc = parent = getter = undefined; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
 
 var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
 
@@ -38,6 +38,9 @@ var invariant = require('./invariant');
 var GraphQLRelayDirective = {
   name: 'relay',
   args: [{
+    name: 'pattern',
+    type: types.GraphQLBoolean
+  }, {
     name: 'plural',
     type: types.GraphQLBoolean
   }]
@@ -103,6 +106,11 @@ var RelayQLNode = (function () {
         return new RelayQLDirective(_this2.context, directive);
       });
     }
+  }, {
+    key: 'isPattern',
+    value: function isPattern() {
+      return this.context.isPattern;
+    }
   }]);
 
   return RelayQLNode;
@@ -133,7 +141,13 @@ var RelayQLFragment = (function (_RelayQLDefinition) {
   function RelayQLFragment(context, ast, parentType) {
     _classCallCheck(this, RelayQLFragment);
 
-    _get(Object.getPrototypeOf(RelayQLFragment.prototype), 'constructor', this).call(this, context, ast);
+    // @relay(pattern: true)
+    var isPattern = (ast.directives || []).some(function (directive) {
+      return directive.name.value === 'relay' && (directive.arguments || []).some(function (arg) {
+        return arg.name.value === 'pattern' && arg.value.kind === 'BooleanValue' && arg.value.value;
+      });
+    });
+    _get(Object.getPrototypeOf(RelayQLFragment.prototype), 'constructor', this).call(this, _extends({}, context, { isPattern: isPattern }), ast);
     this.parentType = parentType;
   }