+++ scripts/babel-relay-plugin/src/RelayQLTransformer.js
@@ -153,6 +153,7 @@ class RelayQLTransformer {
 
     const context = {
       definitionName: capitalize(documentName),
+      isPattern: false,
       schema: this.schema,
     };
     if (definition.kind === 'FragmentDefinition') {