+++ src/traversal/__tests__/filterRelayQuery-test.js
@@ -29,7 +29,7 @@ describe('filterRelayQuery()', () => {
     query = getNode(Relay.QL`
       query {
         viewer {
-          newsFeed {
+          newsFeed(first: "10") {
             edges {
               node {
                 message {
@@ -63,7 +63,7 @@ describe('filterRelayQuery()', () => {
     expect(filterRelayQuery(query, filter)).toEqualQueryRoot(getNode(Relay.QL`
       query {
         viewer {
-          newsFeed {
+          newsFeed(first: "10") {
             edges {
               cursor,
               node {