+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## master
 
+* Bump the `babel-relay-plugin` version to vX.Y.Z.
+  * Added validation of arguments for connections with `edges` or `pageInfo`.
+    * Connections without arguments in fat queries can add the new fragment
+      directive `@relay(pattern: true)`.
+  * Fixed validation of fields within inline fragments in connections.
 * `RelayQueryField#getStorageKey` will now produce the same key regardless of
   the order of a field's arguments.
 * Range behavior keys in mutation configs are now guaranteed to be sorted.